Updating tests and fixing some code smells from SonarrCloud
