+++ Sources/SearchOps/Keychain/KeychainOperations.swift
@@ -9,7 +9,13 @@ import Foundation
 
 public class KeychainOperations : KeychainOperationsProtocol {
   
-  public init(){}
+  public init() {
+    // Required to create a public initialiser
+    // For a type that’s defined as public, the default initializer is considered internal.
+    // If you want a public type to be initializable with a no-argument initializer when used in another module, you must explicitly provide a public no-argument initializer yourself as part of the type’s definition.
+    // https://docs.swift.org/swift-book/documentation/the-swift-programming-language/accesscontrol/
+    // SonarCloud - swift:S1186
+  }
   
   public func SecItemCopyMatching(query:[String : Any]) throws -> Data {
     var item: AnyObject? = nil
